Enumerable k-nearest-neighbour search functionality (#259)

+++ jdbc-adapter/src/main/java/org/polypheny/db/adapter/jdbc/stores/PostgresqlStore.java
@@ -150,7 +150,7 @@ public class PostgresqlStore extends AbstractJdbcStore {
             case VARBINARY:
                 return "VARBINARY";
             case TINYINT:
-                return "TINYINT";
+                return "SMALLINT";
             case SMALLINT:
                 return "SMALLINT";
             case INTEGER:
